requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-ssm
aws-cdk.aws-ec2
aws-cdk.aws-kms
aws-cdk.aws-iam
aws-cdk.aws-rds
aws-cdk.aws-secretsmanager
aws-cdk.aws-elasticache
aws-cdk.aws-cognito
aws-cdk.aws-apigateway
aws-cdk.aws-lambda
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-codebuild
aws-cdk.aws-events-targets
aws-cdk.aws-sns
aws-cdk.aws-sns-subscriptions
aws-cdk.aws-cloudfront
aws-cdk.aws-wafv2
aws-cdk.aws-route53
aws-cdk.aws-route53-targets
aws-cdk.aws-cloudtrail
aws-cdk.aws-elasticsearch
aws-cdk.aws-eks
